// src/modules/party/dto/acbs-get-party-by-search-text-response-element.dto.ts
export interface AcbsGetPartyBySearchTextResponseElement {
  PartyAlternateIdentifier: string;
  IndustryClassification: { IndustryClassificationCode: string };
  PartyName1: string;
  PartyName2: string;
  PartyName3: string;
  MinorityClass: { MinorityClassCode: string };
  CitizenshipClass: { CitizenshipClassCode: string };
  OfficerRiskDate: string;
  PrimaryAddress: { Country: { CountryCode: string } };
}